Add try; Add --dry to install.py and making dirs

+++ install.py
@@ -1,5 +1,6 @@
 #!/usr/bin/python
 
+import sys
 import os
 import pathlib
 
@@ -9,22 +10,35 @@ files = [
     '.Xresources',
     '.config/i3/config',
     '.config/i3status/config',
-    '.vim/ycm_extra_conf.py'
+    '.vim/ycm_extra_conf.py',
+    '.try/cpp',
+    '.try/cpp_src',
+    '.bin/try'
 ]
+
+dry = False
+
+for i in range(1, len(sys.argv)):
+    arg = sys.argv[i]
+    if arg == '--dry':
+        dry = True
+    else:
+        print('unknown argument \'' + arg + '\'')
+        exit(1)
 
 def source(f):
     return os.path.join( os.path.dirname( os.path.realpath(__file__) ), f )
-
-home_dir = os.path.expanduser('~')
-backup_dir = source('backup')
-print('using home directory \'' + home_dir + '\'')
-print('using backup directory: \'' + backup_dir + '\'')
 
 def home(f):
     return os.path.join(home_dir, f)
 
 def backup(f):
     return os.path.join(backup_dir, f)
+
+home_dir = os.path.expanduser('~')
+backup_dir = source('backup')
+print('using home directory \'' + home_dir + '\'')
+print('using backup directory: \'' + backup_dir + '\'')
 
 def install():
     # backup existing files
@@ -35,20 +49,30 @@ def install():
 
         if os.path.exists(dst):
             if os.path.islink(dst):
-                print ('removing existing link \'' + dst + '\'')
-                os.remove(dst)
+                print('removing existing link \'' + dst + '\'')
+                if not dry:
+                    os.remove(dst)
             else:
-                print ('backing up \'' + dst + '\' -> \'' + bak + '\'')
+                print('backing up \'' + dst + '\' -> \'' + bak + '\'')
                 pathlib.Path(os.path.dirname(bak)).mkdir(parents=True, exist_ok=True)
-                os.rename(dst, bak)
+                if not dry:
+                    os.rename(dst, bak)
+
+        # create directory if doesnt exist
+        dst_dir = os.path.dirname(dst)
+        if not os.path.exists(dst_dir):
+            print('creating directory \'' + dst_dir + '\'')
+            if not dry:
+                os.makedirs(dst_dir)
 
     # create symbolic links
     for f in files:
         src = source(f)
         dst = home(f)
 
-        print ('creating symbolic link \'' + src + '\' -> \'' + dst + '\'')
-        os.symlink(src, dst)
+        print('creating symbolic link \'' + src + '\' -> \'' + dst + '\'')
+        if not dry:
+            os.symlink(src, dst)
 
 def uninstall():
     pass
